+++ crates/primitives/src/signature/mod.rs
@@ -1,0 +1,11 @@
+mod error;
+pub use error::SignatureError;
+
+mod parity;
+pub use parity::Parity;
+
+mod sig;
+pub(crate) use sig::Signature;
+
+mod utils;
+pub use utils::to_eip155_v;
